gl: Decode textures in parallel
This is a great speedup on multicore computers, as we can decode
multiple textures in parallel, and also do the GLFW setup while textures
are being decoded.

This uses a (too?) trivial dispatch algorithm: each new texture is sent
to the next thread, looping back to the first one once each one has been
sent one texture.  It doesn’t take into account whether a thread has
already received work or how fast it was to complete it, which could
prove sub-optimal.