Add ASTC support
Just like for BC7 we have to include a header to describe the texture,
here I chose to copy the 16-byte ASTC header as is, before doing the
Zstd compression on the actual data.

This format is suitable for use on modern mobile phones, and can allow
some huge savings in bandwidth, both network and memory.  For instance a
4096×4096 texture can get encoded from 16 MiB (using 4×4 blocks, same as
BC7) down to 1.9 MiB (using 12×12 blocks).