wasm: Add web assembly support
This works in browsers, using WebGL.

I haven’t found out a way to tell cargo to use a specific crate-type
when the target_arch is wasm32, so for now you will have to add this to
the Cargo.toml:
[lib]
crate-type = ["cdylib"]

The size of the canvas and browser zoom is read once at startup, then
never again, it doesn’t react dynamically to the size of the canvas
changing.

Aside from that, all features from the native GLFW version are
available.